VacationDAO 인터페이스 추가, VacationMapper NAMESPACE 변경
