+++ src/main/resources/templates/vacation/request.html
@@ -90,7 +90,7 @@
                     <tr>
                         <td class="section-title" style="height: 80px">휴가 사유</td>
                         <td>
-                            <textarea id="reason" name="reason" rows="5" placeholder="휴가 사유를 입력해주세요."></textarea>
+                            <textarea id="reason" name="reason" rows="5" placeholder="휴가 사유를 입력해주세요." maxlength="255"></textarea>
                         </td>
                     </tr>
                 </table>
